Add Exercism data to the docs
This adds the Exercism curriculum with all of the language tracks that have implementations
of at least one of the exercises. Each language track is in a separate repository, making it
much easier to manage.

This does not include the contributing information for the main website
and the command-line client, as at the moment we're trying to get the
governance worked out for the language tracks and general curriculum
and not as focused on the site (which has other issues).